src/backend/requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-openai
langchain-community
numpy
networkx
exa_py
langchain_experimental
spacy
python-dotenv
beautifulsoup4
trafilatura
lxml
thefuzz
requests
sentence-transformers
faiss-cpu
pydantic
aiohttp
scipy
scikit-learn
census
sqlalchemy
redis
supabase
boto3
pyathena
census
requests
httpx

